+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+4.67.2
+==========
+### Bug Fixes
+- Fixed an issue with Managed Identity source detection where Azure ML was prioritized incorrectly over Azure App Service, causing token acquisition failures. See [Bug #5077](https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/5077)
+  
+4.67.1
+==========
+### Bug Fixes
+- Added support for Azure Machine Learning (Azure ML) Managed Identity (MSI) to address missing functionality for token acquisition in Azure Machine Learning environments. See [Bug #4984](https://github.com/AzureAD/microsoft-authentication-library-for-dotnet/issues/4984)
+
 4.67.0
 ==========
 ### New Features